refactor: フロアマップ機能もcomponentsディレクトリを作成しIndex.tsを用いて機能ごとのcomponentsを取り出しやすくした
