+++ src/app/roomHistory/page.tsx
@@ -1,7 +1,7 @@
 'use client';
-import RoomHistory from '@/features/roomHistory/RoomHistory';
+import * as RoomHistoryComponents from '@/features/roomHistory/components/Index';
 
 const RoomHistoryPage = () => {
-  return <RoomHistory />;
+  return <RoomHistoryComponents.RoomHistory />;
 };
 export default RoomHistoryPage;
