+++ src/features/floorMap/components/Index.ts
@@ -1,0 +1,5 @@
+import { FloorMapCanvas } from './FloorMapCanvas';
+import { PopoverTop } from './PopoverTop';
+import { PopoverUser } from './PopoverUser';
+
+export { FloorMapCanvas, PopoverTop, PopoverUser };
