+++ src/features/floorMap/components/PopoverUser.tsx
@@ -4,7 +4,7 @@ type Props = {
   userNames: string[];
 };
 
-const PopoverUser = ({ roomName, userNames }: Props) => {
+export const PopoverUser = ({ roomName, userNames }: Props) => {
   return (
     <div className='mt-2 min-w-fit whitespace-nowrap border-2 border-gray-500 bg-white p-[5%] shadow-md'>
       <p className='text-gray-500'>{roomName}</p>
@@ -14,5 +14,3 @@ const PopoverUser = ({ roomName, userNames }: Props) => {
     </div>
   );
 };
-
-export default PopoverUser;
