+++ src/features/floorMap/hooks/useFloorMapCanvas.tsx
